Refactor refreshTaskList function to improve code maintainability and solved error when deleting a task when it is the only task on the list
