+++ main.js
@@ -2,6 +2,7 @@ import { Task } from './js/models/Task.js';
 import { getTasks, saveTasks } from './js/utils/tasks-management.js';
 import { setTaskListMaxHeight } from './js/utils/calc-tasklist-height.js';
 import { labelTopFormAnimation } from './js/animations/label-top-animation.js';
+import { refreshTaskList } from './js/utils/refresh-tasklist.js';
 
 // Get the main components of the page
 const taskList = document.querySelector(".tasklist");
@@ -9,73 +10,6 @@ const actionButtons = taskList.querySelectorAll("button");
 const deleteAllTasksButton = actionButtons[0];
 const addNewTaskButton = actionButtons[1];
 const logoutButton = document.querySelector(".user-info button");
-const taskListUl = taskList.querySelector("ul");
-
-// Function to display the tasks on the page
-function refreshTaskList() {
-
-  // Clear the task list
-  taskListUl.innerHTML = "";
-
-  setTaskListMaxHeight();
-
-  // Get the tasks from the local storage or an empty array
-  // TODO: Filter the tasks by the user
-  const tasks = getTasks();
-  const tasksNumber = tasks.length;
-  
-  // If there are no tasks...
-  if (tasksNumber === 0) {
-    deleteAllTasksButton.disabled = true; // Disable the button
-    const noTasksp = document.createElement("p");
-    noTasksp.innerText = "No tasks to display";
-    noTasksp.id = "no-tasks";
-    taskList.insertBefore(noTasksp, addNewTaskButton);
-    return;
-  }
-
-  // Enable the delete all tasks button
-  deleteAllTasksButton.disabled = false;
-
-  // Delete the 'No tasks to display' message if it exists
-  const taskListp = taskList.querySelector("#no-tasks");
-  if (taskListp) {
-    taskList.removeChild(taskListp);
-  }
-
-
-  // Create a new unordered list
-  tasks.forEach((task) => {
-    const taskTemplate = document.getElementById('task-template');
-    const clone = taskTemplate.content.cloneNode(true);
-    const title = clone.querySelector("h3")
-    const p_elements = clone.querySelectorAll("p")
-    const description = p_elements[0];
-    const dueDate = p_elements[1];
-    const deleteButton = clone.querySelector("button");
-
-    // TODO: To correctly delete the task, it is needed to validate that the title is unique
-    deleteButton.onclick = () => {
-      // Get the tasks from the local storage or an empty array
-      const tasks = getTasks();
-
-      // Filter the tasks to remove the task with the title of the current task
-      const newTasks = tasks.filter((t) => t.title !== task.title);
-
-      // Save the new tasks array to the local storage
-      saveTasks(newTasks);
-
-      // Refresh the task list
-      refreshTaskList();
-    };
-
-    title.innerText = `Title: ${task.title}`;
-    description.innerText = `Description: ${task.description}`;
-    dueDate.innerText = `Due date: ${task.dueDate}`;
-
-    taskListUl.append(clone);
-  });
-}
 
 // Event listener to adjust the task list max-height when the window is resized
 window.addEventListener("resize", setTaskListMaxHeight);
